implement stream hasher setup
add a new PublicKeyStream type that supports streaming the data to the
internal blake2 hasher. this is beneficial in situations where the data
is processed on the fly and then discareded (e.g. a stream that gets
decompressed to disk on the go).

because of the inherent nature of the stream setup, this new method can
not support the legacy minisign mode of operation where files are not
prehashed.